+++ views/emails/default.blade.php
@@ -1,5 +1,5 @@
 <p>
-    Dear {{name}}},
-    Your email is: {{email}}
+    Dear {{$name}},
+    Your email is: {{$email}}
     -- Sent from crumbform
 </p>