+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 All notable changes to `crumbform` will be documented in this file
 
-## 1.0.0 - XXXX-XX-XX
+## 1.0.0-beta.2 - 2021-07-07
+
+- Fix template bug with variables
+- Change default logging to "file"
+- Update illuminate/support version
+
+## 1.0.0-beta.1 - 2020-12-21
 
 - Logging: Database and file
 - Form validation